Extract class to a new file
